(#5) SamplePage を pages 以下に移動する
